test(unit): split act tests in several files
